app-report/requirements.txt
# FastAPI 프레임워크
fastapi==0.115.12

# ASGI 서버
uvicorn[standard]==0.34.2

# Pydantic 모델링
python-dotenv==1.1.0
pydantic==2.11.4

# logger
psutil==5.9.8
GPUtil==1.4.0

# MCP
langchain-mcp-adapters==0.1.1
langgraph==0.4.5
langchain-ollama==0.3.3
langchain-community==0.3.24
langchain-openai==0.3.1
